Fix error when trying to load an empty table
