+++ bids2table/helpers.py
@@ -2,6 +2,7 @@ from pathlib import Path
 from typing import Any, Dict, Optional, Union
 
 import pandas as pd
+from elbow.typing import StrOrPath
 
 from bids2table.extractors.entities import BIDSEntities
 
@@ -35,8 +36,13 @@ def flat_to_multi_columns(df: pd.DataFrame, sep: str = "__") -> pd.DataFrame:
     if isinstance(df.columns, pd.MultiIndex):
         return df
 
-    columns = df.columns
-    split_columns = [col.split(sep) for col in columns]
+    # Do nothing for empty df
+    # TODO: It would probably be better if the header was initialized even if there are
+    # no records.
+    if len(df.columns) == 0:
+        return df
+
+    split_columns = [col.split(sep) for col in df.columns]
     num_levels = max(map(len, split_columns))
 
     def _pad_col(col):
@@ -60,4 +66,16 @@ def multi_to_flat_columns(df: pd.DataFrame, sep: str = "__") -> pd.DataFrame:
 
     df = df.copy(deep=False)
     df.columns = pd.Index(join_columns)
+    return df
+
+
+def load_index(
+    path: StrOrPath, split_columns: bool = True, sep: str = "__"
+) -> pd.DataFrame:
+    """
+    Load a bids2table index, optionally splitting columns into a multi index on `sep`.
+    """
+    df = pd.read_parquet(path)
+    if split_columns:
+        df = flat_to_multi_columns(df, sep=sep)
     return df
